refactor: improve linux signal handling with documented syscall wrappers, struct ABI clarifications, and updated CI tooling configurations.

+++ crates/varta-watch/src/peer_cred/recv.rs
@@ -182,7 +182,7 @@ pub(crate) fn recv_authenticated(fd: i32) -> RecvResult {
                     _ => return RecvResult::IoError(err),
                 }
             }
-            break ret as isize;
+            break ret;
         };
 
         if plat::ctrl_truncated(&mhdr) {
